+++ packages/@sanity/vanilla-extract-integration/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @sanity/vanilla-extract-integration
 
+## 0.1.6
+
+### Patch Changes
+
+- [#3113](https://github.com/sanity-io/pkg-utils/pull/3113) [`dceee04`](https://github.com/sanity-io/pkg-utils/commit/dceee045e2115f89784ba01acd8a6861ec8da579) Thanks [@squiggler-app](https://github.com/apps/squiggler-app)! - fix(deps): update dependency tsdown to ^0.22.13
+
+- [#3140](https://github.com/sanity-io/pkg-utils/pull/3140) [`d37e63c`](https://github.com/sanity-io/pkg-utils/commit/d37e63cf2af8cab20a493aadde08a30a9a927f2c) Thanks [@squiggler-app](https://github.com/apps/squiggler-app)! - fix(deps): update dependency yuku-parser to ^0.8.0
+
 ## 0.1.5
 
 ### Patch Changes